+++ crate/src/models/core_taxon.rs
@@ -28,7 +28,7 @@ pub struct CoreTaxon {
     #[serde(rename = "rank", skip_serializing_if = "Option::is_none")]
     pub rank: Option<String>,
     #[serde(rename = "rank_level", skip_serializing_if = "Option::is_none")]
-    pub rank_level: Option<i32>,
+    pub rank_level: Option<f32>,
 }
 
 impl CoreTaxon {
